# openapi.yaml
openapi: 3.0.0
info:
  title: Todo API
  version: 1.0.0
paths:
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created
        '400':
          description: Invalid input
  /api/auth/login:
    post:
      summary: Login and obtain JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: {type: string}
                  token_type: {type: string}
  /api/tasks:
    get:
      summary: List all tasks for authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of tasks
    post:
      summary: Create a new task
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: {type: string}
                description: {type: string}
      responses:
        '201':
          description: Task created
  /api/tasks/{task_id}:
    put:
      summary: Update a task (mark as completed)
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                completed: {type: boolean}
      responses:
        '200':
          description: Task updated
    delete:
      summary: Delete a task
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Task deleted